requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.6.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0
